LPS-105035 Reorganizing the upgrade path to be able to upgrade from any version without repeating any upgrade process
